BL-16822 Extract the canvas control menu rendering from the canvas component
CanvasElementContextControls owned both the decision of which controls to show and
the rendering of the menu that shows them. Inline (Word-style) images need the second
without the first: the picture's menu is the canvas element menu, item for item, but
the picture is not a canvas element and the canvas element manager knows nothing
about it.

So the rendering moves to its own module, canvasControlMenuRendering.tsx, and the
canvas component calls it. No behaviour changes and no inline-image code appears
here; this is the seam the rest of the stack builds on.

Two small things the same need turned up:

- getImageContainer treated an element whose img is a direct child, with no
  bloom-imageContainer, as having no container. That is the shape of a legacy image
  and of an inline image, and it is already the fallback
  buildCanvasElementControlRegistryContext uses for hasImage, so the two now agree.
- IControlContext gains an optional deleteThisObject, for an object that is not a
  canvas element and whose deletion means more than removing one element. Left out,
  Delete removes the page's active canvas element as it always has.
