+++ src/BloomBrowserUI/bookEdit/toolbox/canvas/canvasControlRegistry.ts
@@ -42,7 +42,6 @@ import { default as VolumeUpIcon } from "@mui/icons-material/VolumeUp";
 import { getWorkspaceBundleExports } from "../../js/workspaceFrames";
 import {
     doImageCommand,
-    getImageFromCanvasElement,
     getImageFromContainer,
     getImageTransparencyMode,
     getImageUrlFromImageContainer,
@@ -115,7 +114,12 @@ const getImageContainer = (ctx: IControlContext): HTMLElement | undefined => {
     if (imageContainer) {
         return imageContainer;
     }
-    return getImageFromCanvasElement(ctx.canvasElement)
+    // An element whose img is a direct child, with no bloom-imageContainer -- a legacy
+    // shape, or an inline image in a text block (whose wrapper is the "canvas element"
+    // here) -- is its own container. Same fallback as
+    // buildCanvasElementControlRegistryContext, whose hasImage these commands are
+    // offered against.
+    return getImageFromContainer(ctx.canvasElement)
         ? ctx.canvasElement
         : undefined;
 };
@@ -1255,7 +1259,14 @@ export const controlRegistry: Record<TopLevelControlId, IControlDefinition> = {
         menu: {
             iconScale: 1.2,
         },
-        action: () => {
+        action: (ctx) => {
+            // See IControlContext.deleteThisObject: an inline image deletes itself, because
+            // it is not the page's active canvas element and deleting it means more than
+            // removing one element.
+            if (ctx.deleteThisObject) {
+                ctx.deleteThisObject();
+                return;
+            }
             getCanvasElementManager()?.deleteCurrentCanvasElement?.();
         },
     },
