+++ src/BloomBrowserUI/bookEdit/toolbox/canvas/canvasControlTypes.ts
@@ -155,6 +155,12 @@ export interface IControlContext {
     // edit (see aiImageEditorImageFormats.ts). False for formats the editor can't open
     // (e.g. svg), which keeps "Edit with AI" disabled for them.
     imageIsAiEditableFormat: boolean;
+    // How to delete the object these controls are about, for an object that is not a canvas
+    // element. An inline image (a picture inside a text block) sets this, because deleting
+    // one means taking its copy out of every language's text, which the canvas element
+    // manager knows nothing about. Left out, Delete removes the page's active canvas
+    // element, as it always has.
+    deleteThisObject?: () => void;
 }
 
 export interface IControlRuntime {
